requirements.txt
numpy
tensorflow
Pillow
scikit-learn
opencv-python
Werkzeug
Flask
pylint
black